Day-18/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit>=1.28.0
transformers>=4.30.0
torch>=2.0.0

